fix(search): cast BM25 ts_rank_cd to double precision for consistent scoring
- Eliminates float32→float64 precision artifacts (e.g. 0.85→0.8500000238418579)
- Both BM25 queries now match VectorSearch CAST pattern
- Regenerated sqlc: BM25SearchRow.Score is now float64 natively

+++ internal/server/handlers/bm25.go
@@ -59,7 +59,7 @@ func BM25Search(q BM25SearchQuerier, logger zerolog.Logger) echo.HandlerFunc {
 			Tags          []string
 			CreatedAt     time.Time
 			UpdatedAt     time.Time
-			Score         float32
+			Score         float64
 		}
 
 		var matched []bm25Row
@@ -112,7 +112,7 @@ func BM25Search(q BM25SearchQuerier, logger zerolog.Logger) echo.HandlerFunc {
 				ID:            r.ID,
 				Title:         r.Title,
 				Snippet:       truncateSnippet(r.Content, maxSnippetLen),
-				Score:         float64(r.Score),
+				Score:         r.Score,
 				Tags:          r.Tags,
 				Collection:    r.Collection,
 				WorkspaceHash: r.WorkspaceHash,
